Moved munin/algorithm to munin/detail/algorithm
Since it is only used in components' implementation details.
