+++ src/detail/algorithm.cpp
@@ -1,9 +1,9 @@
-#include "munin/algorithm.hpp"
+#include "munin/detail/algorithm.hpp"
 #include <terminalpp/canvas.hpp>
 #include <terminalpp/canvas_view.hpp>
 #include <algorithm>
 
-namespace munin {
+namespace munin { namespace detail {
 
 // ==========================================================================
 // INTERSECTION
@@ -286,5 +286,5 @@ void copy_region(
     }
 }
 
-}
+}}
